labeling-tool/labeling-tool/backend/requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
starlette==0.38.4
PyJWT==2.8.0
python-multipart==0.0.12
neo4j